// src/app/types/others.ts
import { JSX } from "react";

export interface SearchObj {
  page: number;
  limit: number;
  order: string;
}

export interface ProductSearchObj {
  product_volume?: string;
  page: number;
  limit: number;
  order?: string;
  brand_mb_id?: string;
  product_name?: string;
  product_collection?: string;
  product_size?: string;
  product_color?: string;
  product_type?: string;
  min_price?: number;
  max_price?: number;
  my_favorite?: boolean;
}
export interface ProductSearch {
  page: number;
  limit: number;
  order: string;
  product_name: string;
  brand_mb_id?: string;
  product_collection?: string;
  product_size?: string;
  product_color?: string;
  product_volume?: number;
  product_type?: string;
}

export interface CommentsSearchObj {
  page: number;
  limit: number;
  comment_ref_product_id: string;
  order: string;
}

export interface MemberLiken {
  like_group: string;
  like_status: number;
  like_ref_id: string;
}

export interface CartItem {
  _id: string;
  quantity: number;
  name: string;
  price: number;
  image: string;
}

export interface ChatMessage {
  msg: string;
  mb_id: string;
  mb_nick: string;
  mb_image: string;
}
export interface ChatGreetMsg {
  text: string;
}

export interface ChatInfoMsg {
  total: number;
}

export interface ProductNew {
  page: number;
  limit: number;
  order: string;
  product_collection?: string;
  product_size?: string;
  product_color?: string;
  product_type?: string;
}

export interface NavbarObj {
  element: JSX.Element;
  title: string;
  path: string;
  private: boolean;
  hidden: boolean;
}